Module pearl::filter::bloom

Bloom filter aHash version 0.7.4 (commit ffa04fcb81f39755f636c75c9b7aa06533c0ae75) is used for Bloom filter. Several key tests were performed in order to determine the best hash function for this scenario. Results for aHash were compared with Wyhash as the second best after aHash.

Results are the following:

SMHasher a. AHash: All tests passed b. Wyhash: PerlinNoise test failed

SMHasher3 a. AHash: 118 / 188 passed b. Wyhash: 160 / 188 passed

Mean Bloom filter’s false-positive amount of 100k keys in 1M-sized bitvec: a. AHash: 29574.6 b. Wyhash: 29546

Average Speed: a. AHash: 23.20 cycles/hash b. Wyhash: 22.27 cycles/hash

Structs

Bloom

Bloom filter

Config

Bloom filter configuration parameters.