Struct pearl::Builder

pub struct Builder { /* fields omitted */ }

Is used to initialize a Storage.

Required params:

work_dir - where Storage will keep blob and index files

max_blob_size - upper limit of blob file size

max_data_in_blob - maximum number of records in one blob

blob_file_name_prefix - prefix in blob file name pattern: {prefix}.{id}.{ext}

key_size - const key size in bytes

Example

use pearl::{Builder, Storage, Key};

struct Id;

let storage: Storage<Id> = Builder::new()
        .blob_file_name_prefix("benchmark")
        .max_blob_size(10_000_000)
        .max_data_in_blob(1_000)
        .work_dir(std::env::temp_dir().join("pearl_benchmark"))
        .build()
        .unwrap();

Implementations

impl Builder

#[must_use]
pub fn new() -> Self

Create new uninitialized Builder

pub fn build<K: Key>(self) -> Result<Storage<K>>

Creates Storage based on given configuration, returns error if not all params are set.

Errors

Return error if some of the required params is missed

pub fn work_dir<S: Into<PathBuf>>(self, work_dir: S) -> Self

Sets working directory. If path doesn’t exists, Storage will try to create it at initialization stage.

#[must_use]
pub fn max_blob_size(self, max_blob_size: u64) -> Self

Sets blob file size approximate limit. When the file size exceeds it, active blob update is activated. Must be greater than zero

#[must_use]
pub fn max_data_in_blob(self, max_data_in_blob: u64) -> Self

Limits max number of records in a single blob. Must be greater than zero

pub fn blob_file_name_prefix<U: Into<String>>(
self,
blob_file_name_prefix: U
) -> Self

Sets blob file name prefix, e.g. if prefix set to hellopearl, files will be named as hellopearl.[N].blob. Where N - index number of file If the prefix is empty, param won’t be set.

#[must_use]
pub fn allow_duplicates(self) -> Self

Disables check existence of the record on write. Writing becomes faster because there is no additional disk access for searching for duplicates.

pub fn ignore_corrupted(self) -> Self

Ignores blobs with corrupted header for index file and continues work with normal ones (also leaves error logs for bad blobs). It’s worth noticing that index file maybe corrupted likely in case of irregular stop of the storage: either the disk disconnected or server crashed. In these cases blob file is likely either corrupted, so there is no a try to restore indices from blob file.

#[must_use]
pub fn set_filter_config(self, config: BloomConfig) -> Self

Sets custom bloom filter config, if not set, use default values.

#[must_use]
pub fn enable_aio(self, ioring: Rio) -> Self

Enables linux AIO with provided io_uring.

#[must_use]
pub fn create_work_dir(self, create: bool) -> Self

[Optional] Sets whether to create work directory if its missing on storage initialization. Default value is true

pub fn set_dump_sem(self, dump_sem: Arc<Semaphore>) -> Self

Sets semaphore for index dumping on blob change. Parallel saving of indexes onto the disk will be limited by this semaphore. This can prevent disk overusage in systems with multiple pearls.